svg в историю редактирования продукции
