+++ src/components/CatCharacteristics/Right.js
@@ -1,0 +1,7 @@
+/**
+ *
+ *
+ * @module Right
+ *
+ * Created by Evgeniy Malyarov on 05.07.2021.
+ */
